+++ src/http/token/client_credentials.rs
@@ -129,6 +129,8 @@ pub(crate) async fn token_client_credentials(
             mtls_x5t_s256,
             refresh_token_mtls_x5t_s256: None,
             authorization_code_hash: None,
+            actor: None,
+            issued_token_type: None,
         },
     )
     .await