+++ src/http/token/device.rs
@@ -487,6 +487,8 @@ pub(crate) async fn token_device_code(
                     mtls_x5t_s256: mtls_x5t_s256.clone(),
                     refresh_token_mtls_x5t_s256: mtls_x5t_s256,
                     authorization_code_hash: None,
+                    actor: None,
+                    issued_token_type: None,
                 },
             )
             .await